당첨 금액은 큰 수의 곱으로 이루어지기 때문에 long 타입을 return하도록 수정하였으며, 수익률 계산 시 당첨 금액 / 구입 금액 * 100으로 변경하였습니다. 또한, 당첨 금액이 0인 경우 0으로 나누는 문제가 발생합니다. 이 문제를 해결하기 위해 0인 경우는 바로 0을 리턴하도록 변경하였습니다.

+++ src/main/java/lotto/domain/LottoPrize.java
@@ -61,12 +61,12 @@ public class LottoPrize {
             messageUtil.printEarningRate(earningRate);
         }
 
-        private int getTotalWinningAmount() {
-            return THREE_COUNT.getMatchCount() * THREE_COUNT.getPrice()
-                    + FOUR_COUNT.getMatchCount() * FOUR_COUNT.getPrice()
-                    + FIVE_COUNT.getMatchCount() * FIVE_COUNT.getPrice()
-                    + FIVE_COUNT_WITH_BONUS.getMatchCount() * FIVE_COUNT_WITH_BONUS.getPrice()
-                    + SIX_COUNT.getMatchCount() * SIX_COUNT.getPrice();
+        private long getTotalWinningAmount() {
+            return (long) THREE_COUNT.getMatchCount() * THREE_COUNT.getPrice()
+                    + (long) FOUR_COUNT.getMatchCount() * FOUR_COUNT.getPrice()
+                    + (long) FIVE_COUNT.getMatchCount() * FIVE_COUNT.getPrice()
+                    + (long) FIVE_COUNT_WITH_BONUS.getMatchCount() * FIVE_COUNT_WITH_BONUS.getPrice()
+                    + (long) SIX_COUNT.getMatchCount() * SIX_COUNT.getPrice();
 
         }
     }
